Add the binary directly to the LittleFS. Follow this tutorial: https://randomnerdtutorials.com/arduino-ide-2-install-esp32-littlefs/ Adding the LittleFS uploader for convenience.
Add a flush command to flush the serial at the beginning of the programming.
Try to connect and program in a loop, makes testing easier. Testing with a STM32F411.
Erasing the first page seems to work, but erasing the second fails. Why? Different mapping between L4 and F4 chips?
Programming fails too.